Start implementing reparenting and window drawing

+++ src/style.rs
@@ -2,8 +2,8 @@ use xcb::render::Color;
 
 // TODO implement fonts and colors
 pub struct Style {
-    pub border_width: u8,
-    pub titlebar_height: u8,
+    pub border_width: i16,
+    pub titlebar_height: i16,
 
     pub titlebar_color_bg: Color,
     pub titlebar_color_fg: Color,
